modules/kvs: Move kvs_util convenience library
Move kvs_util convenience library from modules/kvs to common/libkvs.
Rename kvs_util.h to kvs_util_private.h.  Also move kvs_util unit
tests.
