+++ src/common/libkvs/kvs_util.c
