+++ src/modules/kvs/Makefile.am
@@ -19,8 +19,6 @@ kvs_la_SOURCES = \
 	cache.h \
 	waitqueue.c \
 	waitqueue.h \
-	kvs_util.h \
-	kvs_util.c \
 	lookup.h \
 	lookup.c \
 	treq.h \
@@ -42,7 +40,6 @@ TESTS = \
 	test_lookup.t \
 	test_treq.t \
 	test_kvstxn.t \
-	test_kvs_util.t \
 	test_kvsroot.t
 
 test_ldadd = \
@@ -66,7 +63,6 @@ test_waitqueue_t_SOURCES = test/waitqueue.c
 test_waitqueue_t_CPPFLAGS = $(test_cppflags)
 test_waitqueue_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/waitqueue.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
 test_cache_t_SOURCES = test/cache.c
@@ -74,7 +70,6 @@ test_cache_t_CPPFLAGS = $(test_cppflags)
 test_cache_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/cache.o \
 	$(top_builddir)/src/modules/kvs/waitqueue.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
 test_lookup_t_SOURCES = test/lookup.c
@@ -86,14 +81,12 @@ test_lookup_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/kvsroot.o \
 	$(top_builddir)/src/modules/kvs/kvstxn.o \
 	$(top_builddir)/src/modules/kvs/treq.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
 test_treq_t_SOURCES = test/treq.c
 test_treq_t_CPPFLAGS = $(test_cppflags)
 test_treq_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/treq.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
 test_kvstxn_t_SOURCES = test/kvstxn.c
@@ -105,13 +98,6 @@ test_kvstxn_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/kvsroot.o \
 	$(top_builddir)/src/modules/kvs/treq.o \
 	$(top_builddir)/src/modules/kvs/waitqueue.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
-	$(test_ldadd)
-
-test_kvs_util_t_SOURCES = test/kvs_util.c
-test_kvs_util_t_CPPFLAGS = $(test_cppflags)
-test_kvs_util_t_LDADD = \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
 test_kvsroot_t_SOURCES = test/kvsroot.c
@@ -122,5 +108,4 @@ test_kvsroot_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/kvstxn.o \
 	$(top_builddir)/src/modules/kvs/cache.o \
 	$(top_builddir)/src/modules/kvs/treq.o \
-	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
